[Gameplay] Create a generic 'Longsword' mixin that adds a small splash attack and apply it to the Thracian Black Cloak
This mixin is genericized in order to be useful for other units if desired. It adds a splash attack, but removes some regular attack to balance. With this, a champion or other unit can be easily differentiated. In this instance, it is applied to the Macedonians' champion swordsman, the Thracian Black Cloak. These guys use the very long Rhomphaia and Sika swords of Thraco-Dacian origin. 

Patch by: @wowgetoffyourcellphone
Accepted by: @real_tabasco_sauce, @borg- 

Differential revision: https://code.wildfiregames.com/D5145

git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@27871 3db68df2-c116-0410-a063-a993310a9797