[GUI] Remove Roman numerals from bot names
To avoid the appearance of two Roman numerals at the end of bot names (e.g. "Artaxshasha II II -> here "Artaxshasha II" was the name itself and the second "II" was added when the name was chosen for the second time), remove the addition of Roman numerals entirely.

Instead, if possible, a new name is chosen for each bot. If not, the suffix "(n)" is added counting the duplicates and warning is displayed. AINames are added to the Iberians, Gauls, and Mauryas bringing their totel to at least 8 each in order avoid any duplicates in regular 8-player games.

Credit to Norse_Harold for changing the condition for reusing names to not use a hardcoded assumption about the maximum number of player slots

Patch by @Vantha
Accepted by @phosit, @wowgetoffyourcellphone

Differential Revision: https://code.wildfiregames.com/D5256

git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@28082 3db68df2-c116-0410-a063-a993310a9797