Don't execute the task when no Future awaits it anymore
Summary:
Most of the times the callback stores a reference to a variable in scope where the `Future` is in. When the scope is left the reference get's dangling. `CancelOrWait` is called in multiple places (mostly destructors) to ensure the callback isn't executed anymore.
This patch deduplicates thous calls to `CancelOrWait`.

Refs: #5874

Comments by: @Stan, @vladislavbelov

Differential Revision: https://code.wildfiregames.com/D5208



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@28128 3db68df2-c116-0410-a063-a993310a9797