Fix removal of email addresses from PO-files
The regex for removing the email address of the last translator from
PO-files contained two bugs causing it not to match:

1. It didn't account for text after the email address and PO-files
   from Transifex contain the year of the last translation there.
2. It expected a second `\\n"` at the end of the line, instead of `\n"`.

This commit fixes both of these issues and simplifies the regexes. It
also ensures that the text after the email address is kept as well.

Patch by: @Dunedan
Accepted by: @Stan

Differential Revision: https://code.wildfiregames.com/D5290



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@28130 3db68df2-c116-0410-a063-a993310a9797