Get a promise when starting a GUIpage
When calling `Engine.PushGuiPage` a promise is returned. The promise is settled when the "child" page is closed. That allows to `await` it inside `async` functions.
Previously the callback was run right inside the call to `Engine.PopGuiPage`. Now the continuation of the promise is called at the end of the "tick".

This won't help performance. It will more likely make things worse. Since gui pages aren't opened or closed that frequently, it doesn't matter that much.

Refs: r22676

For the engine side:
The promise is stored in the `CGUIManager::SGUIPage` (like previously the callback). When the promise is fulfilled it enqueues a callback in the `JobQueue` of the `JSContext`.

Comments by: @wraitii, @Stan, @Polakrity, @lyv, @elexis, @vladislavbelov

Differential Revision: https://code.wildfiregames.com/D3807



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@28145 3db68df2-c116-0410-a063-a993310a9797