Preserve comments in debug translations
Up to now the "debug" and "long" translations got generated without the
comments present in the portable object files. This lead to the problem
that "dennis-ignore" directives in the comments weren't present in these
translations anymore, causing dennis to report false-positives for the
debug translations. This commit fixes that by keeping the comments in
the debug translations.

Patch by: @Dunedan
Accepted by: @Stan

Differential Revision: https://code.wildfiregames.com/D5303


git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@28150 3db68df2-c116-0410-a063-a993310a9797