Revert "Custom context for specific names"
r28160 had some unintended consequences. Some specific names apparently
are indeed supposed to be translated. Also using the message context
the translator hint isn't the right place and changes the primary key
used for these strings on Transifex. Instead comments should be used.
To avoid more interruption to translators, we revert the changes for
now and come up with a better patch.


git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@28162 3db68df2-c116-0410-a063-a993310a9797