Improve formatting of XMPP announcements
Previously private messages and announcements got handled the same way.
This adds a distinct handling for server announcements, which allows
formatting them in a different way. They are now not prefixed with
"Private" anymore and show the announcement subject as well, if one was
set.

Patch by: @Dunedan
Accepted by: @Stan

Differential Revision: https://code.wildfiregames.com/D5268


git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@28186 3db68df2-c116-0410-a063-a993310a9797