Make messageBox return a promise
Ofthen the `messageBox` is only used for confirmation like: "Do you really want to overwrite the savegame?"
With this patch the code paths of "overwrite savegame" and "make new savegame" can be combined.

In some cases there is nothing done when one button is pressed. That is changed to an early return.

Comments by: @Stan

Differential Revision: https://code.wildfiregames.com/D5274



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@28191 3db68df2-c116-0410-a063-a993310a9797